Split withdrawal hook tests
